Updated MO logic to postgres + fixed defense progress bar for longer defense campaigns + changes to targetName for MO killtasks

+++ server/plugins/initDB.ts
@@ -4,49 +4,49 @@ export default defineNitroPlugin(() => {
 
   try {
   
-  sql`CREATE TABLE IF NOT EXISTS orders
-  
-  (
-      id32 INTEGER PRIMARY KEY,
-      expiresIn INTEGER,
-      brief TEXT,
-      description TEXT,
-      tasks TEXT,
-      progress TEXT,
-      isActive INTEGER
-  )
-  
-  `
-  
-  sql`CREATE TABLE IF NOT EXISTS activePlanets
-  
-  (
-      planetIndex INTEGER PRIMARY KEY,
-      name TEXT,
-      faction TEXT,
-      players INTEGER,
-      health INTEGER,
-      maxHealth INTEGER,
-      percentage NUMERIC,
-      defense INTEGER,
-      majorOrder INTEGER,
-      biome TEXT,
-      expireDateTime NUMERIC
-  )
-  
-  `
-  
-  sql`CREATE TABLE IF NOT EXISTS planets
-  
-  (
-      id INTEGER PRIMARY KEY,
-      name TEXT,
-      biome TEXT,
-      environmentals TEXT,
-      owner TEXT
-  )
-  
-  `
+    sql`CREATE TABLE IF NOT EXISTS orders
+    
+    (
+        id32 BIGINT PRIMARY KEY,
+        expiresIn INTEGER,
+        brief TEXT,
+        description TEXT,
+        tasks TEXT,
+        progress TEXT,
+        isActive INTEGER
+    )
+    
+    `
+    
+    sql`CREATE TABLE IF NOT EXISTS activePlanets
+    
+    (
+        planetIndex INTEGER PRIMARY KEY,
+        name TEXT,
+        faction TEXT,
+        players INTEGER,
+        health INTEGER,
+        maxHealth INTEGER,
+        percentage NUMERIC,
+        defense INTEGER,
+        majorOrder INTEGER,
+        biome TEXT,
+        expireDateTime NUMERIC
+    )
+    
+    `
+    
+    sql`CREATE TABLE IF NOT EXISTS planets
+    
+    (
+        id INTEGER PRIMARY KEY,
+        name TEXT,
+        biome TEXT,
+        environmentals TEXT,
+        owner TEXT
+    )
+    
+    `
 
   } catch (e) {
     console.error(e, " Failed to initialise database")
